6l: fix warning, setmalloctag declaration, missing header type cases

+++ sys/src/cmd/6l/compat.c
@@ -50,6 +50,6 @@ mysbrk(ulong size)
 }
 
 void
-setmalloctag(void*, ulong)
+setmalloctag(void*, uintptr)
 {
 }
